# Don't use client connection pooling for the distributed cache service