--- a/enterprise/server/util/distributed_client/distributed_client.go
+++ b/enterprise/server/util/distributed_client/distributed_client.go
@@ -43,7 +43,7 @@ type Proxy struct {
 	writeBufPool          *bytebufferpool.VariableWriteBufPool
 	mu                    *sync.Mutex
 	server                *grpc.Server
-	clients               map[string]*grpc_client.ClientConnPool
+	clients               map[string]dcpb.DistributedCacheClient
 	heartbeatCallback     func(ctx context.Context, peer string)
 	hintedHandoffCallback func(ctx context.Context, peer string, r *rspb.ResourceName)
 	listenAddr            string
@@ -60,7 +60,7 @@ func New(env environment.Env, c interfaces.Cache, listenAddr string) *Proxy {
 		listenAddr:   listenAddr,
 		mu:           &sync.Mutex{},
 		// server goes here
-		clients: make(map[string]*grpc_client.ClientConnPool),
+		clients: make(map[string]dcpb.DistributedCacheClient),
 	}
 	if zone := resources.GetZone(); zone != "" {
 		proxy.zone = zone
@@ -131,19 +131,16 @@ func (c *Proxy) getClient(ctx context.Context, peer string) (dcpb.DistributedCac
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	if client, ok := c.clients[peer]; ok {
-		conn, err := client.GetReadyConnection()
-		if err != nil {
-			return nil, status.UnavailableErrorf("no connections to peer %q are ready", peer)
-		}
-		return dcpb.NewDistributedCacheClient(conn), nil
+		return client, nil
 	}
 	log.Debugf("Creating new client for peer: %q", peer)
-	conn, err := grpc_client.DialInternal(c.env, "grpc://"+peer)
+	conn, err := grpc_client.DialInternalWithoutPooling(c.env, "grpc://"+peer)
 	if err != nil {
 		return nil, err
 	}
-	c.clients[peer] = conn
-	return dcpb.NewDistributedCacheClient(conn), nil
+	client := dcpb.NewDistributedCacheClient(conn)
+	c.clients[peer] = client
+	return client, nil
 }
 
 func (c *Proxy) prepareContext(ctx context.Context) context.Context {