These connections don't go through a proxy or load balancer, and our servers support unlimited streams per connection, so we just need 1 client connection to each other app.