@vanja-p 2025-08-13
Slightly surprising, but this also makes the benchmarks faster:
```
                                      │     base      │               nopool                │
                                      │    sec/op     │   sec/op     vs base                │
Set/DistPebble10/AC-24                   170.4µ ±  3%   164.7µ ± 2%   -3.35% (p=0.001 n=11)
Set/DistPebble10/CAS-24                  145.3µ ±  5%   135.3µ ± 2%   -6.82% (p=0.000 n=11)
Set/DistPebble100/AC-24                  184.0µ ±  3%   171.6µ ± 1%   -6.72% (p=0.000 n=11)
Set/DistPebble100/CAS-24                 137.9µ ± 13%   138.2µ ± 1%        ~ (p=0.898 n=11)
Set/DistPebble1000/AC-24                 174.3µ ± 20%   191.1µ ± 2%        ~ (p=0.300 n=11)
Set/DistPebble1000/CAS-24                141.4µ ±  3%   136.8µ ± 2%   -3.23% (p=0.005 n=11)
Set/DistPebble10000/AC-24                401.5µ ±  8%   362.2µ ± 1%   -9.78% (p=0.000 n=11)
Set/DistPebble10000/CAS-24               160.5µ ±  1%   144.5µ ± 2%  -10.00% (p=0.000 n=11)
Set/DistPebble1000000/AC-24              4.677m ± 15%   4.561m ± 1%        ~ (p=0.065 n=11)
Set/DistPebble1000000/CAS-24            1093.7µ ± 12%   976.6µ ± 7%  -10.71% (p=0.004 n=11)
Set/LookasideDistPebble10/AC-24          189.9µ ±  8%   168.6µ ± 1%  -11.20% (p=0.000 n=11)
Set/LookasideDistPebble10/CAS-24         147.4µ ±  1%   140.3µ ± 1%   -4.79% (p=0.000 n=11)
Set/LookasideDistPebble100/AC-24         188.3µ ± 16%   182.4µ ± 2%   -3.13% (p=0.003 n=11)
Set/LookasideDistPebble100/CAS-24        144.6µ ±  1%   142.0µ ± 1%   -1.79% (p=0.000 n=11)
Set/LookasideDistPebble1000/AC-24        211.6µ ±  7%   198.7µ ± 1%   -6.09% (p=0.005 n=11)
Set/LookasideDistPebble1000/CAS-24       161.8µ ±  7%   143.3µ ± 1%  -11.43% (p=0.001 n=11)
Set/LookasideDistPebble10000/AC-24       377.7µ ±  2%   357.7µ ± 2%   -5.29% (p=0.000 n=11)
Set/LookasideDistPebble10000/CAS-24      159.2µ ±  4%   148.3µ ± 2%   -6.85% (p=0.000 n=11)
Set/LookasideDistPebble1000000/AC-24     4.474m ±  1%   4.481m ± 1%        ~ (p=0.365 n=11)
Set/LookasideDistPebble1000000/CAS-24   1388.5µ ± 33%   988.8µ ± 6%  -28.79% (p=0.002 n=11)
geomean                                  305.4µ         285.8µ        -6.41%

                                      │     base      │                 nopool                 │
                                      │      B/s      │     B/s       vs base                  │
Set/DistPebble10/AC-24                  58.59Ki ±  0%   58.59Ki ± 0%        ~ (p=1.000 n=11)
Set/DistPebble10/CAS-24                 68.36Ki ±  0%   68.36Ki ± 0%        ~ (p=0.262 n=11)
Set/DistPebble100/AC-24                 527.3Ki ±  7%   566.4Ki ± 2%   +7.41% (p=0.000 n=11)
Set/DistPebble100/CAS-24                712.9Ki ± 12%   703.1Ki ± 1%        ~ (p=0.626 n=11)
Set/DistPebble1000/AC-24                5.474Mi ± 21%   4.988Mi ± 2%        ~ (p=0.291 n=11)
Set/DistPebble1000/CAS-24               6.742Mi ±  6%   6.971Mi ± 2%   +3.39% (p=0.004 n=11)
Set/DistPebble10000/AC-24               23.76Mi ±  8%   26.33Mi ± 1%  +10.84% (p=0.000 n=11)
Set/DistPebble10000/CAS-24              59.42Mi ±  3%   66.02Mi ± 2%  +11.11% (p=0.000 n=11)
Set/DistPebble1000000/AC-24             203.9Mi ± 20%   209.1Mi ± 1%        ~ (p=0.065 n=11)
Set/DistPebble1000000/CAS-24            872.0Mi ± 13%   976.6Mi ± 6%  +12.00% (p=0.004 n=11)
Set/LookasideDistPebble10/AC-24         48.83Ki ± 20%   58.59Ki ± 0%  +20.00% (p=0.012 n=11)
Set/LookasideDistPebble10/CAS-24        68.36Ki ±  0%   68.36Ki ± 0%        ~ (p=1.000 n=11) ¹
Set/LookasideDistPebble100/AC-24        517.6Ki ± 13%   537.1Ki ± 2%   +3.77% (p=0.001 n=11)
Set/LookasideDistPebble100/CAS-24       673.8Ki ±  1%   683.6Ki ± 1%   +1.45% (p=0.002 n=11)
Set/LookasideDistPebble1000/AC-24       4.511Mi ±  7%   4.797Mi ± 1%   +6.34% (p=0.005 n=11)
Set/LookasideDistPebble1000/CAS-24      5.894Mi ±  8%   6.657Mi ± 2%  +12.94% (p=0.000 n=11)
Set/LookasideDistPebble10000/AC-24      25.25Mi ±  3%   26.66Mi ± 3%   +5.59% (p=0.000 n=11)
Set/LookasideDistPebble10000/CAS-24     59.89Mi ±  6%   64.30Mi ± 2%   +7.36% (p=0.000 n=11)
Set/LookasideDistPebble1000000/AC-24    213.1Mi ±  1%   212.8Mi ± 1%        ~ (p=0.365 n=11)
Set/LookasideDistPebble1000000/CAS-24   686.8Mi ± 32%   964.5Mi ± 4%  +40.42% (p=0.002 n=11)
geomean                                 4.953Mi         5.266Mi        +6.32%
¹ all samples are equal

                                      │     base     │               nopool                │
                                      │     B/op     │     B/op      vs base               │
Set/DistPebble10/AC-24                  44.59Ki ± 0%   44.56Ki ± 0%  -0.07% (p=0.016 n=11)
Set/DistPebble10/CAS-24                 36.72Ki ± 0%   36.69Ki ± 0%  -0.09% (p=0.001 n=11)
Set/DistPebble100/AC-24                 44.98Ki ± 0%   44.96Ki ± 0%       ~ (p=0.193 n=11)
Set/DistPebble100/CAS-24                37.12Ki ± 0%   37.08Ki ± 0%  -0.10% (p=0.023 n=11)
Set/DistPebble1000/AC-24                46.88Ki ± 0%   46.83Ki ± 0%       ~ (p=0.101 n=11)
Set/DistPebble1000/CAS-24               37.78Ki ± 0%   37.72Ki ± 0%  -0.16% (p=0.004 n=11)
Set/DistPebble10000/AC-24               57.38Ki ± 0%   57.32Ki ± 0%       ~ (p=0.151 n=11)
Set/DistPebble10000/CAS-24              46.67Ki ± 0%   46.60Ki ± 0%  -0.16% (p=0.000 n=11)
Set/DistPebble1000000/AC-24             1.074Mi ± 1%   1.071Mi ± 0%       ~ (p=0.365 n=11)
Set/DistPebble1000000/CAS-24            1.027Mi ± 0%   1.028Mi ± 0%       ~ (p=0.699 n=11)
Set/LookasideDistPebble10/AC-24         44.32Ki ± 0%   44.26Ki ± 0%  -0.13% (p=0.004 n=11)
Set/LookasideDistPebble10/CAS-24        36.52Ki ± 0%   36.47Ki ± 0%  -0.14% (p=0.000 n=11)
Set/LookasideDistPebble100/AC-24        44.71Ki ± 0%   44.66Ki ± 0%  -0.10% (p=0.000 n=11)
Set/LookasideDistPebble100/CAS-24       36.89Ki ± 0%   36.85Ki ± 0%  -0.10% (p=0.002 n=11)
Set/LookasideDistPebble1000/AC-24       46.57Ki ± 0%   46.52Ki ± 0%  -0.12% (p=0.008 n=11)
Set/LookasideDistPebble1000/CAS-24      37.54Ki ± 0%   37.49Ki ± 0%  -0.14% (p=0.000 n=11)
Set/LookasideDistPebble10000/AC-24      56.90Ki ± 0%   56.81Ki ± 0%  -0.16% (p=0.001 n=11)
Set/LookasideDistPebble10000/CAS-24     46.30Ki ± 0%   46.26Ki ± 0%  -0.10% (p=0.000 n=11)
Set/LookasideDistPebble1000000/AC-24    1.071Mi ± 1%   1.069Mi ± 0%       ~ (p=0.193 n=11)
Set/LookasideDistPebble1000000/CAS-24   1.034Mi ± 1%   1.028Mi ± 0%  -0.63% (p=0.003 n=11)
geomean                                 82.52Ki        82.40Ki       -0.14%

                                      │    base    │              nopool               │
                                      │ allocs/op  │ allocs/op   vs base               │
Set/DistPebble10/AC-24                  628.0 ± 0%   627.0 ± 0%  -0.16% (p=0.000 n=11)
Set/DistPebble10/CAS-24                 511.0 ± 0%   510.0 ± 0%  -0.20% (p=0.000 n=11)
Set/DistPebble100/AC-24                 626.0 ± 0%   625.0 ± 0%  -0.16% (p=0.000 n=11)
Set/DistPebble100/CAS-24                511.0 ± 0%   510.0 ± 0%  -0.20% (p=0.000 n=11)
Set/DistPebble1000/AC-24                622.0 ± 0%   620.0 ± 0%  -0.32% (p=0.000 n=11)
Set/DistPebble1000/CAS-24               507.0 ± 0%   505.0 ± 0%  -0.39% (p=0.000 n=11)
Set/DistPebble10000/AC-24               658.0 ± 0%   656.0 ± 0%  -0.30% (p=0.000 n=11)
Set/DistPebble10000/CAS-24              508.0 ± 0%   506.0 ± 0%  -0.39% (p=0.000 n=11)
Set/DistPebble1000000/AC-24             716.0 ± 1%   708.0 ± 0%  -1.12% (p=0.000 n=11)
Set/DistPebble1000000/CAS-24            574.0 ± 1%   569.0 ± 0%  -0.87% (p=0.000 n=11)
Set/LookasideDistPebble10/AC-24         628.0 ± 0%   626.0 ± 0%  -0.32% (p=0.000 n=11)
Set/LookasideDistPebble10/CAS-24        511.0 ± 0%   510.0 ± 0%  -0.20% (p=0.000 n=11)
Set/LookasideDistPebble100/AC-24        626.0 ± 0%   625.0 ± 0%  -0.16% (p=0.000 n=11)
Set/LookasideDistPebble100/CAS-24       511.0 ± 0%   510.0 ± 0%  -0.20% (p=0.000 n=11)
Set/LookasideDistPebble1000/AC-24       622.0 ± 0%   620.0 ± 0%  -0.32% (p=0.000 n=11)
Set/LookasideDistPebble1000/CAS-24      507.0 ± 0%   505.0 ± 0%  -0.39% (p=0.000 n=11)
Set/LookasideDistPebble10000/AC-24      657.0 ± 0%   656.0 ± 0%  -0.15% (p=0.000 n=11)
Set/LookasideDistPebble10000/CAS-24     507.0 ± 0%   506.0 ± 0%  -0.20% (p=0.000 n=11)
Set/LookasideDistPebble1000000/AC-24    715.0 ± 0%   703.0 ± 0%  -1.68% (p=0.000 n=11)
Set/LookasideDistPebble1000000/CAS-24   579.0 ± 2%   564.0 ± 0%  -2.59% (p=0.000 n=11)
geomean                                 582.0        578.9       -0.52%
```